infra/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "basec-cli"
version = "0.1.0"
description = "CLI profissional para gerenciar infraestrutura BaseCommerce"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "paramiko>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
basec = "basec.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["basec*"]

